# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WhisperBox API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getPublicProfile
      tags: [users]
      summary: Get public profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "404":
          description: User not found
...
  /campaigns/me:
    get:
      operationId: getMyCampaign
      tags: [campaigns]
      summary: Get my active QnA campaign
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active campaign or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "204":
          description: No active campaign
        "401":
          description: Unauthorized

    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new QnA campaign (Premium only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "401":
          description: Unauthorized
        "403":
          description: Premium required
  /campaigns/{id}/end:
    patch:
      operationId: endCampaign
      tags: [campaigns]
      summary: End (deactivate) a campaign
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "401":
          description: Unauthorized
        "404":
          description: Campaign not found
  /campaigns/public/{username}:
    get:
      operationId: getPublicCampaign
      tags: [campaigns]
      summary: Get active campaign for a public profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "204":
          description: No active campaign
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Campaign:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        question:
          type: string
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        responseCount:
          type: integer
      required:
        - id
        - userId
        - title
        - question
        - isActive
        - createdAt
        - responseCount
    CreateCampaignBody:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        question:
          type: string
          maxLength: 500
        color:
          type: string
        icon:
          type: string
      required:
        - title
        - question
...
    PublicProfile:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        socialInstagram:
          type: string
          nullable: true
        socialTiktok:
          type: string
          nullable: true
        socialX:
          type: string
          nullable: true
        socialFacebook:
          type: string
          nullable: true
        socialGithub:
          type: string
          nullable: true
        socialLinkedin:
          type: string
          nullable: true
        isPremium:
          type: boolean
        allowReplyNotif:
          type: boolean
        publicMessages:
          type: array
          items:
            $ref: "#/components/schemas/PublicMessage"
      required:
        - username
        - isPremium
        - allowReplyNotif
        - publicMessages
    PublicMessage:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        ownerReply:
          type: string
          nullable: true
        ownerRepliedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - content
        - createdAt
...